Implement requested changes.
 - Using ordinal & slicce in MC-8187
 - Call a common method rather than calling into a mixin method
 - Separating out the test suites for each bug
 - Rename Utilities to Utils
 - Omit flwr8187$ or any other modid prefix.
